chatbot-backend/requirements.txt
# chatbot-backend/requirements.txt

fastapi
uvicorn[standard]
python-dotenv
groq 
sentence-transformers
faiss-cpu
pymongo
python-multipart